Eliminate need for intermediate files.
Code in try.sh has been tested; ready to test shrink.sh.

+++ shrink.sh
@@ -65,39 +65,54 @@ RES1="py1.tmp"
 RES2="py2.tmp"
 
 # Part ONE: use dumpe2fs to get the size of our 2nd partition:
-echo "Use dumpe2fs to get block count and size of 2nd partition..."
-dumpe2fs -h ${LOOPDEV}p2 | grep -e "Block count" -e "Block size" > $RES
+##echo "Use dumpe2fs to get block count and size of 2nd partition..."
+##dumpe2fs -h ${LOOPDEV}p2 | grep -e "Block count" -e "Block size" > $RES
 # The result goes into a temporary file to be parsed...
 # ...by the next command: 
-./getblockcount.py $RES $RES1 $RES2
+##./getblockcount.py $RES $RES1 $RES2
 # Takes as input content of the $HRES file and puts the block count
 # and the block size into the $RES1 and $RES2 files respectively.
 
 # We now collect those values into integer variables:
-exec 3< $RES1; read LINE <&3; declare -i COUNT=$LINE
-exec 4< $RES2; read LINE <&4; declare -i SIZE=$LINE
+##exec 3< $RES1; read LINE <&3; declare -i COUNT=$LINE
+##exec 4< $RES2; read LINE <&4; declare -i SIZE=$LINE
+
+# Part ONE (V2):
+declare -a DATA=`dumpe2fs -h $PARTITION | \
+    grep -e "Block count" -e "Block size" | \
+    ./getblockcount.py`
+declare -i COUNT=${DATA[0]}
+declare -i SIZE=${DATA[1]}
 
 # Progress report:
 printf "Count (sh) is %d, and size is %d.\n" $COUNT $SIZE
 
 
 # Part TWO: get the starting blocks of each of the partitions:
-echo "fdisk -l to collect partition data..."
-fdisk -l $LOOPDEV > $RES
+##echo "fdisk -l to collect partition data..."
+##fdisk -l $LOOPDEV > $RES
 # The results are again dumped into a temporary file
 # to be parsed by the next command:
-./getstartsectors.py $RES $RES1 $RES2
+##./getstartsectors.py $RES $RES1 $RES2
 
 # As before, We collect those values into integer variables:
-exec 5< $RES1; read LINE <&5; declare -i P1_1stSECTOR=$LINE
-exec 6< $RES2; read LINE <&6; declare -i P2_1stSECTOR=$LINE
+##exec 5< $RES1; read LINE <&5; declare -i P1_1stSECTOR=$LINE
+##exec 6< $RES2; read LINE <&6; declare -i P2_1stSECTOR=$LINE
+
+# Part TWO (V2): get the starting blocks of each of the partitions:
+echo "fdisk -l to list partitions..."
+declare -a DATA=`fdisk -l $DEV | ./getstartsectors.py`
+declare -i P1_1stSECTOR=${DATA[0]}
+declare -i P2_1stSECTOR=${DATA[1]}
 
 printf "Beginning (sh) sectors are %d and %d.\n" \
 P1_1stSECTOR $P2_1stSECTOR
 echo "We don't need the first sector of the first partition."
 
 # We no longer have need of our temporary files...
-rm $RES $RES1 $RES2  # ...unless we are still debugging.
+if test -f $RES ; then
+    rm $RES $RES1 $RES2  # ...unless we are still debugging.
+fi
 
 # So now we have
 # 1. The block count of the second partition &
